refactor(editor): extract first-diff auto-scroll from DiffViewer
Pure code move, no behavior change. DiffViewer is near the 400-line
ceiling and the next feature would push it over; this effect is
self-contained and reads better named.
